Added fieldgroups to init and display
Pulled field display into a separate function, so both fieldgroups and boxes can call it to loop.

Changed switch fieldtype default.
Ideally would like to remove the switch completely and let all displays happen via the filter!
changed the earlier default to case 'text', since we have alreday set it to that for type-less fields
defined the new default with a "slt_cf_custom_display_filter" filter and passed all the required display fields to this filter
(there has to be a better way of doing this, will revisit)

Changed init heavily!
Defined a new function "slt_cf_init_fieldtypes" where the default field types and their defaults are stored as arrays
This function contains the filter "register_new_dc_fields" that passes all the fields to the filter. So, now default settings and fields can be modified using this filter
A new field type can be added or an existing one removed

Created two new functions to filter the above by default, for file and gmap fields. a lot of work remains here.

Modified the way defaults are checked and validated, slightly so that new fields can be plugged in, without rewriting much code. This has to be made smoother

Signed-off-by: Saurabh Shukla <saurabh@yapapaya.com>